backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
azure-cosmos==4.5.1
azure-storage-blob==12.19.0
azure-identity==1.15.0
python-dotenv==1.0.0
